requirements.txt
biopython==1.81
click==8.1.8
pytest==8.3.4
